requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.6
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
numpy==2.0.0
openai==1.35.10
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
